move filedrop handler into separate component use editor container component to collect drop events
